addressed comments
- fixed rubik font problem
- moved files into ui folder
- better responsiveness for existing components and moved headers out of grids
- etc
