+++ apps/hyperdrive-trading/src/globals.css
@@ -1,3 +1,6 @@
+/* Google fonts */
+@import url("https://fonts.googleapis.com/css2?family=Rubik:ital,wght@0,300;0,400;0,500;0,600;0,700;1,300;1,400;1,500;1,600;1,700&display=swap");
+
 /* Custom fonts */
 @font-face {
   font-family: "Akira";
@@ -31,6 +34,3 @@
   font-style: italic;
   font-weight: bold;
 }
-
-/* Google fonts */
-@import url("https://fonts.googleapis.com/css2?family=Rubik&display=swap");
