+++ apps/hyperdrive-trading/src/pages/Markets.tsx
@@ -1,10 +1,10 @@
-import { FeaturedMarkets } from "markets/components/FeaturedMarkets";
-import { ProtocolStatsBar } from "protocol/components/ProtocolStatsBar";
+import { FeaturedMarkets } from "src/markets/components/FeaturedMarkets";
+import { ProtocolStatsBar } from "src/protocol/components/ProtocolStatsBar";
 import { ReactElement } from "react";
 
 export function Markets(): ReactElement {
   return (
-    <div className="m-auto flex max-w-5xl flex-col gap-y-12 px-6 pt-[5vh]">
+    <div className="m-auto flex max-w-5xl flex-col gap-y-12">
       <ProtocolStatsBar />
       <FeaturedMarkets />
     </div>
