+++ apps/hyperdrive-trading/src/protocol/components/ProtocolStatsBar.tsx
@@ -1,34 +1,20 @@
 import { ReactElement } from "react";
+import { Stat } from "src/ui/base/components/Stat";
 
 interface ProtocolStatsBarProps {}
 
 export function ProtocolStatsBar({}: ProtocolStatsBarProps): ReactElement {
   return (
-    <div className="grid grid-cols-2 items-center gap-x-24 gap-y-4 rounded bg-base-200 px-12 py-4 md:grid-cols-protocol-stats">
-      <h2 className="col-span-2 font-quantico text-2xl font-bold text-off-white md:col-span-1">
+    <div className="grid grid-flow-row items-center gap-y-8 gap-x-24 rounded bg-base-200 px-12 py-8 md:grid-flow-col md:py-4">
+      <h2 className="font-quantico text-2xl font-bold text-off-white">
         Protocol
       </h2>
-      <Stat label="TVL" value="$150,000,000" />
-      <Stat label="Volume (7D)" value="$1,000,000" />
-      <Stat label="Fees (7D)" value="$10,000" />
-    </div>
-  );
-}
 
-interface StatProps {
-  label: string;
-  value: string;
-}
-
-function Stat({ label, value }: StatProps): ReactElement {
-  return (
-    <div>
-      <p className="mb-1 whitespace-nowrap font-quantico text-lg text-giga-blue-200">
-        {label}
-      </p>
-      <p className="whitespace-nowrap font-rubik text-lg font-semibold text-off-white">
-        {value}
-      </p>
+      <div className="grid grid-cols-2 items-center gap-x-16 gap-y-4 md:grid-cols-[repeat(4,minmax(70px,110px))]">
+        <Stat label="TVL" value="$150,000,000" />
+        <Stat label="Volume (7D)" value="$1,000,000" />
+        <Stat label="Fees (7D)" value="$10,000" />
+      </div>
     </div>
   );
 }
