+++ apps/hyperdrive-trading/src/ui/base/components/Navbar.tsx
@@ -11,9 +11,9 @@ export function Navbar(): ReactElement {
       <div className="navbar-start">
         <Link
           to="/"
-          className="whitespace-nowrap font-akira text-2xl normal-case text-hyper-pink"
+          className="whitespace-nowrap font-akira text-2xl uppercase text-hyper-pink"
         >
-          HYPERDRIVE TRADING
+          Hyperdrive
         </Link>
       </div>
