+++ apps/hyperdrive-trading/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "@hyperdrive/tsconfig/base.json",
   "compilerOptions": {
-    "baseUrl": "./src",
+    "baseUrl": ".",
     "target": "esnext",
     "lib": ["dom", "dom.iterable", "esnext"],
     "esModuleInterop": true,
